Complete Phase 5d ZK functional slice

+++ org.adempiere.pos/src/main/java/ui/zk/org/adempiere/pos/POSNumberBox.java
@@ -154,7 +154,7 @@ public class POSNumberBox extends Div
     public void setValue(Object value)
     {
     	if (value == null)
-    		decimalBox.setValue(null);
+		decimalBox.setValue((BigDecimal) null);
     	else if (value instanceof BigDecimal)
     		decimalBox.setValue((BigDecimal) value);
     	else if (value instanceof Number)
